Restructure idp command.
