+++ src/cli/idp/idp-export.js
@@ -1,0 +1,90 @@
+import { Command, Option } from 'commander';
+import * as common from '../cmd_common.js';
+import { getTokens } from '../../api/AuthApi.js';
+import {
+  exportProvider,
+  exportProvidersToFile,
+  exportProvidersToFiles,
+} from '../../ops/IdpOps.js';
+import storage from '../../storage/SessionStorage.js';
+import { printMessage } from '../../api/utils/Console.js';
+
+const program = new Command('frodo idp export');
+
+program
+  .description('Export (social) identity providers.')
+  .helpOption('-h, --help', 'Help')
+  .showHelpAfterError()
+  .addArgument(common.hostArgumentM)
+  .addArgument(common.realmArgument)
+  .addArgument(common.userArgument)
+  .addArgument(common.passwordArgument)
+  .addOption(common.deploymentOption)
+  .addOption(common.insecureOption)
+  .addOption(
+    new Option(
+      '-i, --idp-id <idp-id>',
+      'Id/name of a provider. If specified, -a and -A are ignored.'
+    )
+  )
+  .addOption(
+    new Option(
+      '-f, --file [file]',
+      'Name of the file to write the exported provider(s) to. Ignored with -A.'
+    )
+  )
+  .addOption(
+    new Option(
+      '-a, --all',
+      'Export all the providers in a realm to a single file. Ignored with -t and -i.'
+    )
+  )
+  .addOption(
+    new Option(
+      '-A, --all-separate',
+      'Export all the providers in a realm as separate files <provider name>.idp.json. Ignored with -t, -i, and -a.'
+    )
+  )
+  .action(
+    // implement command logic inside action handler
+    async (host, realm, user, password, options) => {
+      storage.session.setTenant(host);
+      storage.session.setRealm(realm);
+      storage.session.setUsername(user);
+      storage.session.setPassword(password);
+      storage.session.setDeploymentType(options.type);
+      storage.session.setAllowInsecureConnection(options.insecure);
+      if (await getTokens()) {
+        // export by id/name
+        if (options.idpId) {
+          printMessage(
+            `Exporting provider "${
+              options.idpId
+            }" from realm "${storage.session.getRealm()}"...`
+          );
+          exportProvider(options.idpId, options.file);
+        }
+        // --all -a
+        else if (options.all) {
+          printMessage('Exporting all providers to a single file...');
+          exportProvidersToFile(options.file);
+        }
+        // --all-separate -A
+        else if (options.allSeparate) {
+          printMessage('Exporting all providers to separate files...');
+          exportProvidersToFiles();
+        }
+        // unrecognized combination of options or no options
+        else {
+          printMessage(
+            'Unrecognized combination of options or no options...',
+            'error'
+          );
+          program.help();
+        }
+      }
+    }
+    // end command logic inside action handler
+  );
+
+program.parse();
